server/requirements.txt
Flask==3.1.2
Flask_Caching==2.3.1
Flask-Cors==6.0.1
Flask-Limiter==3.12
pydantic==2.11.7
pyswisseph==2.10.3.2
